feat: add web interface for code generation field properties
